feat: split the config builder to the proper package\
